Align all elements of the container.
I had to put a z-index on the .search div otherwise the box-shadow would
not show up. A lot of tinkering went into getting things close to the
model given. I had to ignore white space wrapping on a header, give
negative margins to elements, and overall find a way to make it similar.
At this point I feel like I may enjoy javascript over css but we will
see.